team3ProjectSourcing/Python-Backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
Pillow
playwright
